enable Werror

+++ platform/msm_shared/crypto4_eng.c
@@ -218,7 +218,7 @@ crypto_send_data(void *ctx_ptr, unsigned char *data_ptr,
 		       buff_ptr,
 		       (((buff_size + sha1_ctx->saved_buff_indx) <=
 			 CRYPTO_SHA_BLOCK_SIZE)
-			? buff_size : (CRYPTO_SHA_BLOCK_SIZE -
+			? buff_size : (unsigned int)(CRYPTO_SHA_BLOCK_SIZE -
 				       sha1_ctx->saved_buff_indx)));
 
 		if (bytes_to_write >= CRYPTO_SHA_BLOCK_SIZE) {
